+++ design-system/tokens/examples/constellation/token-modal.tsx
@@ -5,7 +5,7 @@ import Button from '@atlaskit/button/standard-button';
 import { N0, N30A, N60A } from '@atlaskit/theme/colors';
 
 // eslint-disable-next-line @atlassian/tangerine/import/no-relative-package-imports
-import { Example } from '../../../../../services/website-constellation/src/__DO_NOT_ADD_TO_THIS_FOLDER__/gatsby-theme-brisk/components/example/Example';
+import { Example } from '../../../../../services/design-system-docs/src/__DO_NOT_ADD_TO_THIS_FOLDER__/gatsby-theme-brisk/components/example/Example';
 import token from '../../src/get-token';
 
 const TokensModalCodeBlock = `