+++ design-system/tokens/examples/constellation/token-table.tsx
@@ -12,7 +12,7 @@ import DynamicTable from '@atlaskit/dynamic-table';
 import { N50A, N60A } from '@atlaskit/theme/colors';
 
 // eslint-disable-next-line @atlassian/tangerine/import/no-relative-package-imports
-import { Example } from '../../../../../services/website-constellation/src/__DO_NOT_ADD_TO_THIS_FOLDER__/gatsby-theme-brisk/components/example/Example';
+import { Example } from '../../../../../services/design-system-docs/src/__DO_NOT_ADD_TO_THIS_FOLDER__/gatsby-theme-brisk/components/example/Example';
 import token from '../../src/get-token';
 
 const TokensTagCodeBlock = `