+++ design-system/tokens/examples/constellation/token-tag.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { SimpleTag as Tag } from '@atlaskit/tag';
 
 // eslint-disable-next-line @atlassian/tangerine/import/no-relative-package-imports
-import { Example } from '../../../../../services/website-constellation/src/__DO_NOT_ADD_TO_THIS_FOLDER__/gatsby-theme-brisk/components/example/Example';
+import { Example } from '../../../../../services/design-system-docs/src/__DO_NOT_ADD_TO_THIS_FOLDER__/gatsby-theme-brisk/components/example/Example';
 
 const TokensTagCodeBlock = `
 import { N800, P100, P500, P75 } from '@atlaskit/theme/colors';