+++ dev-portal/rating/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/rating
 
+## 0.2.38
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.2.37
 
 ### Patch Changes